% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epcot.R
\docType{package}
\name{epcot}
\alias{epcot}
\title{epcot: Exome-sequencing Paired Case-Only Tests}
\description{
Analysis package accompanying premature birth Danish sister pairs whole exome sequencing project (PI: Ryckman).
}
