% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epcot_example}
\alias{epcot_example}
\title{Example data to illustrate/test the functionality of epcot}
\format{
A list with three elements, each containing the 1000 rows (one for each
variant:
\describe{
  \item{N2}{Counts for sister pairs}
  \item{N3}{Counts for sister trios}
  \item{vData}{Metadata for each variant}
}
}
\usage{
epcot_example
}
\description{
A random sample of 1000 variants from clb/ryckman/wes/data/rds, N-pair.rds,
N-trio.rds, and vData.rds
}
\keyword{datasets}
